perf: 5.9x faster self-play, behaviour unchanged
Profiling showed the cost was almost entirely representation overhead, not
algorithms: numpy used as a scalar container at sizes where boxing dominates,
and static board data recomputed per call.

Four changes:

1. JSettlers production tables (53% of that workload). _node_production and
   _player_production rescanned all 19 hexes x 6 corners on every call, and
   _node_production did a numpy membership scan per hex. Both are now derived
   once per board; player production is memoised against the ownership bytes,
   and the board-only part of _setup_node_score is precomputed.

2. Python mirrors of the frozen topology tables. Hot paths index these
   19/54/72-entry tables one element at a time, where numpy scalar boxing costs
   far more than it saves. Also cache generate_topology(): it rebuilt the whole
   board graph for every game despite being deterministic.

3. Drop numpy from 5-element vector math. Measured: np.all(hand >= take) is
   1547ns vs 235ns in Python; hand.sum() is 841ns vs 40ns. playable_moves also
   boxed its list into an int32 array every step (926ns) that every caller
   immediately unboxed again -- it now returns a plain list.

4. Plain ints for scalar state fields. (np.uint8(3)+1)%4 is 256ns; (3+1)%4 is
   4ns. prompt, current_player_idx, robber_hex and the piece counts were all
   read as Python ints downstream anyway.

Measured (24 cores for the batch figure):

    jsettlers self-play    11.6 -> 69.2 games/s   5.9x
    engine-only picker     14.1 -> 31.9 games/s   2.3x
    run_batch, 3000 games   111 ->   542 games/s  4.9x

Verified identical, not just plausible: tools/difftest.py runs both trees off
one seeded picker and compares every GameState field after every action -- 200
games, no divergence. tools/regression.py carries golden digests recorded from
main; they pass unchanged here.

+++ catan/init.py
@@ -57,7 +57,7 @@ def initialize_game(rng: np.random.Generator) -> GameState:
         road_owner=np.full(n_road, -1, dtype=np.int8),
         settlement_owner=np.full(n_sett, -1, dtype=np.int8),
         settlement_type=np.full(n_sett, EMPTY, dtype=np.uint8),
-        robber_hex=np.uint16(robber_hex),
+        robber_hex=robber_hex,
         dev_deck=dev_deck,
         bank_res=np.full(N_RES, BANK_STOCK, dtype=np.int16),
     )
@@ -65,15 +65,15 @@ def initialize_game(rng: np.random.Generator) -> GameState:
     players = [
         PlayerState(
             hand=np.zeros(N_RES, dtype=np.int8),
-            ports_mask=np.uint8(0),
+            ports_mask=0,
             dev_cards=np.zeros(5, dtype=np.int8),
             new_dev_cards=np.zeros(5, dtype=np.int8),
-            used_knights=np.uint8(0),
-            longest_road_len=np.uint8(0),
-            discards_required=np.uint8(0),
-            roads_built=np.uint8(0),
-            settlements_built=np.uint8(0),
-            cities_built=np.uint8(0),
+            used_knights=0,
+            longest_road_len=0,
+            discards_required=0,
+            roads_built=0,
+            settlements_built=0,
+            cities_built=0,
         )
         for _ in range(N_PLAYERS)
     ]
